requirements.txt
openai-whisper
pandas

